move actual metrics to MetricsImpl
